+++ src/main/java/fuzs/betteranimationscollection/client/renderer/entity/model/JigglySlimeModel.java
@@ -21,11 +21,12 @@ public class JigglySlimeModel<T extends Entity> extends SlimeModel<T> {
     public JigglySlimeModel(int slimeBodyTexOffY) {
         
         super(slimeBodyTexOffY);
+
+        this.isSlimeGelLayer = slimeBodyTexOffY == 0;
         this.cube = ModelUtil.getAtIndex(super.parts().iterator(), 0);
         this.eye0 = ModelUtil.getAtIndex(super.parts().iterator(), 1);
         this.eye1 = ModelUtil.getAtIndex(super.parts().iterator(), 2);
         this.mouth = ModelUtil.getAtIndex(super.parts().iterator(), 3);
-        this.isSlimeGelLayer = slimeBodyTexOffY == 0;
     }
 
     @Override